refactor(web): move createAppQueryClient out of the providers component file
react-doctor only-export-components: non-component exports in a
component file break Fast Refresh state preservation. The factory now
lives in lib/query-client.ts.
